feat(fs): Phase 2 - FAT32 driver, journal CRC32, flash optimizations
- Update journal to use CRC32 checksums instead of simple rolling hash
- Add FAT32 filesystem driver for secondary disk access
  - Full read/write support with short filename handling
  - FAT chain management and cluster allocation
  - FSInfo tracking for free space
- Add flash storage optimizations to ViperFS
  - Lazy superblock sync (only write when dirty)
  - Sorted block writes for better sequential I/O
  - Flash mode flag for reduced wear

+++ viperdos/kernel/fs/viperfs/journal.hpp
@@ -178,9 +178,9 @@ class Journal {
     mutable Spinlock txn_lock_;
 
     /**
-     * @brief Calculate simple checksum for block data.
+     * @brief Calculate CRC32 checksum for block data.
      */
-    u64 checksum(const void *data, usize len);
+    u32 checksum(const void *data, usize len);
 
     /**
      * @brief Read the journal header from disk.
